Remove executable code from sourced files
After `qconfig` determines the platform is Linux, it tries to determine
the OS flavor.  This was previously running as part of a `source`.
Change this by wrapping those commands in a function.  Call this new
function from the `main` function so that code is only being executed in
one place.